Add refactor class
